+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 Added
   * AZ-424: Add Service Tags (AzureRM v2.42+)
+  * AZ-374: Allow disabling Application Insights deployment.
 
 # v4.1.0 - 2021-01-15